refactor : Change location and folder name
