+++ src/main/java/com/levelUpToast/levelUpToast/auth/domain/Authority.java
@@ -1,0 +1,3 @@
+package com.levelUpToast.levelUpToast.auth.domain;
+
+public enum Authority { MEM , ADMIN }
