+++ src/main/java/com/levelUpToast/levelUpToast/auth/service/memberServiceInf/check/UserIntegrityVerification.java
@@ -1,0 +1,10 @@
+package com.levelUpToast.levelUpToast.auth.service.memberServiceInf.check;
+
+import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
+import com.levelUpToast.levelUpToast.auth.domain.Member;
+
+public interface UserIntegrityVerification {
+    Boolean isIDPresent(String id) throws LevelUpToastEx;
+
+    Member isMember(String requestSeq) throws LevelUpToastEx;
+}
