+++ src/main/java/com/levelUpToast/levelUpToast/auth/service/memberServiceInf/login/LogoutMember.java
@@ -1,0 +1,7 @@
+package com.levelUpToast.levelUpToast.auth.service.memberServiceInf.login;
+
+import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
+
+public interface LogoutMember {
+    void logout(String token) throws LevelUpToastEx ;
+}
