+++ src/main/java/com/levelUpToast/levelUpToast/home/service/SimpleHomeService.java
@@ -3,11 +3,10 @@ package com.levelUpToast.levelUpToast.home.service;
 import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
 import com.levelUpToast.levelUpToast.home.repository.HomeRepository;
 import com.levelUpToast.levelUpToast.home.service.homeServiceInf.HomeService;
-import com.levelUpToast.levelUpToast.member.domain.Member;
-import com.levelUpToast.levelUpToast.member.service.memberServiceInf.check.UserIntegrityVerification;
+import com.levelUpToast.levelUpToast.auth.domain.Member;
+import com.levelUpToast.levelUpToast.auth.service.memberServiceInf.check.UserIntegrityVerification;
 import com.levelUpToast.levelUpToast.product.repository.productRepositoryInf.ProductRepository;
 import com.levelUpToast.levelUpToast.util.utilInf.adapter.product.ProductAdapter;
-import com.levelUpToast.levelUpToast.util.utilInf.recommendation.Recommendation;
 import com.levelUpToast.levelUpToast.bodyForm.requestForm.product.ProductListResponseForm;
 import com.levelUpToast.levelUpToast.product.domain.data.tag.Tag;
 import lombok.RequiredArgsConstructor;
