--- src/main/java/com/levelUpToast/levelUpToast/member/domain/Authority.java
@@ -1,3 +1,0 @@
-package com.levelUpToast.levelUpToast.member.domain;
-
-public enum Authority { MEM , ADMIN }
