--- src/main/java/com/levelUpToast/levelUpToast/member/service/memberServiceInf/check/MemberIsID.java
@@ -1,7 +1,0 @@
-package com.levelUpToast.levelUpToast.member.service.memberServiceInf.check;
-
-import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
-
-public interface MemberIsID {
-    Boolean isMemberID(String id) throws LevelUpToastEx;
-}
