--- src/main/java/com/levelUpToast/levelUpToast/member/service/memberServiceInf/login/LogoutMember.java
@@ -1,7 +1,0 @@
-package com.levelUpToast.levelUpToast.member.service.memberServiceInf.login;
-
-import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
-
-public interface LogoutMember {
-    void logout(String token) throws LevelUpToastEx ;
-}
