--- src/main/java/com/levelUpToast/levelUpToast/member/service/memberServiceInf/signUp/SignUpCreate.java
@@ -1,8 +1,0 @@
-package com.levelUpToast.levelUpToast.member.service.memberServiceInf.signUp;
-
-import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
-import com.levelUpToast.levelUpToast.member.domain.Member;
-
-public interface SignUpCreate {
-    void createMember(Member member) throws LevelUpToastEx;
-}
