--- src/main/java/com/levelUpToast/levelUpToast/member/service/memberServiceInf/signUp/SignUpService.java
@@ -1,10 +1,0 @@
-package com.levelUpToast.levelUpToast.member.service.memberServiceInf.signUp;
-
-import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
-import com.levelUpToast.levelUpToast.bodyForm.requestForm.signUp.SignUpRequestForm;
-
-public interface SignUpService {
-
-    void signUp(SignUpRequestForm form) throws LevelUpToastEx;
-
-}
