+++ src/main/java/com/levelUpToast/levelUpToast/util/memebercheck/MemberIsCheckImpl.java
@@ -2,10 +2,10 @@ package com.levelUpToast.levelUpToast.util.memebercheck;
 
 import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
 import com.levelUpToast.levelUpToast.config.token.tokenManager.tokenManagerInf.TokenManager;
-import com.levelUpToast.levelUpToast.member.service.memberServiceInf.check.MemberIsCheck;
+import com.levelUpToast.levelUpToast.auth.service.memberServiceInf.check.MemberIsCheck;
 import com.levelUpToast.levelUpToast.vendor.service.vendor.VendorService;
-import com.levelUpToast.levelUpToast.member.domain.Member;
-import com.levelUpToast.levelUpToast.member.repository.memberRepositoryInf.MemberRepository;
+import com.levelUpToast.levelUpToast.auth.domain.Member;
+import com.levelUpToast.levelUpToast.auth.repository.memberRepositoryInf.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
